feat: Integrate playing history and update UI
- Implement playing history feature with pagination.
- Add playing history data models, RDS, repository, and use case.
- Update profile cubit and state to handle playing history.
- Modify `stock_selection_cubit` to correctly map stock roles.
- Enhance `completed_item_widget` and `completed_details_screen` UI.
- Update `battleground_page` to correctly display normal stocks.
- Refactor `contest_details_page` with shimmer loading.
- Adjust `battleground_item_widget` layout.
- Use PUT instead of POST for client team updates.

+++ lib/features/home_page/data/models/live_stock_model.dart
@@ -1,7 +1,7 @@
 class LiveStock {
   final String stockId;
   final String symbol;
-  final int points;
+  final double points;
   final double currentPrice;
   final double netChange;
   final bool isPredictionCorrect;
@@ -25,7 +25,7 @@ class LiveStock {
     return LiveStock(
       stockId: json['stockId'],
       symbol: json['symbol'],
-      points: json['points'],
+      points: json['points'].toDouble(),
       currentPrice: json['currentPrice'].toDouble(),
       netChange: json['netChange'].toDouble(),
       isPredictionCorrect: json['isPredictionCorrect'],
@@ -52,7 +52,7 @@ class LiveStock {
 
 class ScoreUpdatePayload {
   final String userTeamId;
-  final int totalPoints;
+  final double totalPoints;
   final int rank;
   final List<LiveStock> liveStocks;
